Ofer: 12.1 14:55 f

+++ src/Protocol.cpp
@@ -276,11 +276,15 @@ void Protocol::process_keyboard(std::string &msg) {
                 break;
             }
             case RETURN: {
-                myDB.remove_book_from_Inv(vector_for_input.at(2), vector_for_input.at(1)); //remove from my inv
-                loaner_name = myDB.get_loaner_name(vector_for_input.at(2));
-                //loaner_name = myDB.getBorrowedMap().at(vector_for_input.at(2));
-                myDB.remove_from_borrowdMap(vector_for_input.at(2));//remove borrower from borrow map
-                send(vector_for_input.at(1), "Returning " + vector_for_input.at(2) + " to " + loaner_name);
+                if(vector_for_input.size()<3)
+                    std::cout << "invalid input from keyboard at return case\n" << std::endl;
+                else{
+                    myDB.remove_book_from_Inv(vector_for_input.at(2), vector_for_input.at(1)); //remove from my inv
+                    loaner_name = myDB.get_loaner_name(vector_for_input.at(2));
+                    //loaner_name = myDB.getBorrowedMap().at(vector_for_input.at(2));
+                    myDB.remove_from_borrowdMap(vector_for_input.at(2));//remove borrower from borrow map
+                    send(vector_for_input.at(1), "Returning " + vector_for_input.at(2) + " to " + loaner_name);
+                }
                 break;
             }
             case STATUS: {
